Reorganize durable function code for clarity
Jira ticket: CAMS-421
