+++ backend/function-apps/dataflows/dataflows.ts
@@ -1,7 +1,15 @@
-import { importDataflowSetup } from './import/import-dataflow';
-// import { migrationSetup } from './migration/migration';
+import { setupExportAndLoadCase } from './import/export-and-load-case';
+import { setupMigrateCases } from './import/migrate-cases';
+import { setupStoreCasesRuntimeState } from './import/store-cases-runtime-state';
+import { setupSyncCases } from './import/sync-cases';
 
-importDataflowSetup();
+setupExportAndLoadCase();
+setupStoreCasesRuntimeState();
+setupSyncCases();
+
+// This can be disabled/removed once migration is complete.
+setupMigrateCases();
 
 // TODO: Renable setup so migrations can be executed.
+// import { migrationSetup } from './migration/migration';
 // migrationSetup();
