+++ backend/lib/use-cases/dataflows/dataflow-types.ts
@@ -20,3 +20,20 @@ export type CaseSyncResults = {
   events: CaseSyncEvent[];
   lastTxId?: string;
 };
+
+/**
+ * getDefaultSummary
+ */
+export function getDefaultSummary(
+  override: Partial<ExportCaseChangeEventsSummary> = {},
+): ExportCaseChangeEventsSummary {
+  return {
+    changedCases: 0,
+    exportedAndLoaded: 0,
+    errors: 0,
+    noResult: 0,
+    completed: 0,
+    faulted: 0,
+    ...override,
+  };
+}
